+++ python/GafferImageTest/__init__.py
@@ -81,7 +81,6 @@ from .FormatPlugTest import FormatPlugTest
 from .OffsetTest import OffsetTest
 from .BlurTest import BlurTest
 from .TextTest import TextTest
-from .OpenColorIOTransformTest import OpenColorIOTransformTest
 from .VectorWarpTest import VectorWarpTest
 from .MirrorTest import MirrorTest
 from .CopyChannelsTest import CopyChannelsTest